requirements.txt
Pillow>=9.0.0
mysql-connector-python>=8.0.0
python-dotenv>=0.15.0